Add advanced combining of subtitle lines
Created a new function for combining subtitle lines:
It calculates the number of characters per second for each line, then loops through them to check if the speaking rate is above or below certain thresholds. If out of bounds, it will look to the next and previous text entries, and if other rules are satisfied, it will combine the entry with the the one that will bring it closer to being in bounds.
- Rules are the maximum length of a combined entry's text, and the maximum gap in time between two entries.
- It does two passes, one using more extreme speaking rates, to ensure that those are processed with priority. Otherwise, because of the max text length, there might be less of a chance the extreme entry will be allowed to combine with another.