Add Subtitle Track Remover Tool, Fix Tool Working Dir
Adds new SubtitleTrackRemover.py script to Tools that fetches the subtitle tracks of a video, and allows the user to select which ones to delete. Good if needing to replace a subtitles track with an updated one.

Also fixed issue where Tools scripts weren't seeing the Scripts folder, so added code to set the root dir to path.